refactor: trim_hash moved to common utils instead of having it twice
